+++ src/Pim/Bundle/CatalogBundle/Event/CategoryEvents.php
@@ -12,22 +12,42 @@ namespace Pim\Bundle\CatalogBundle\Event;
 final class CategoryEvents
 {
     /**
-     * This event is thrown before a category is removed.
+     * This event is thrown before a category get removed.
      *
      * The event listener receives an
-     * Symfony\Component\EventDispatcher\GenericEvent instance.
+     * Akeneo\Component\StorageUtils\Event\RemoveEvent instance.
      *
      * @staticvar string
      */
     const PRE_REMOVE_CATEGORY = 'pim_catalog.pre_remove.category';
 
     /**
-     * This event is thrown before a tree is removed.
+     * This event is thrown after an category get removed.
      *
      * The event listener receives an
-     * Symfony\Component\EventDispatcher\GenericEvent instance.
+     * Akeneo\Component\StorageUtils\Event\RemoveEvent instance.
+     *
+     * @staticvar string
+     */
+    const POST_REMOVE_CATEGORY = 'pim_catalog.post_remove.category';
+
+    /**
+     * This event is thrown before a tree get removed.
+     *
+     * The event listener receives an
+     * Akeneo\Component\StorageUtils\Event\RemoveEvent instance.
      *
      * @staticvar string
      */
     const PRE_REMOVE_TREE = 'pim_catalog.pre_remove.tree';
+
+    /**
+     * This event is thrown after an tree get removed.
+     *
+     * The event listener receives an
+     * Akeneo\Component\StorageUtils\Event\RemoveEvent instance.
+     *
+     * @staticvar string
+     */
+    const POST_REMOVE_TREE = 'pim_catalog.post_remove.tree';
 }