+++ src/Pim/Bundle/CatalogBundle/Event/GroupEvents.php
@@ -15,9 +15,19 @@ final class GroupEvents
      * This event is thrown before a group is removed.
      *
      * The event listener receives an
-     * Symfony\Component\EventDispatcher\GenericEvent instance.
+     * Akeneo\Component\StorageUtils\Event\RemoveEvent instance.
      *
      * @staticvar string
      */
     const PRE_REMOVE = 'pim_catalog.pre_remove.group';
+
+    /**
+     * This event is thrown after an group get removed.
+     *
+     * The event listener receives an
+     * Akeneo\Component\StorageUtils\Event\RemoveEvent instance.
+     *
+     * @staticvar string
+     */
+    const POST_REMOVE = 'pim_catalog.post_remove.group';
 }