+++ src/Pim/Bundle/VersioningBundle/UpdateGuesser/VersionableUpdateGuesser.php
@@ -47,8 +47,8 @@ class VersionableUpdateGuesser implements UpdateGuesserInterface
     public function guessUpdates(EntityManager $em, $entity, $action)
     {
         $pendings = array();
-        if ($entity instanceof VersionableInterface
-            || in_array(ClassUtils::getClass($entity), $this->versionableEntities)
+        if ($entity instanceof VersionableInterface ||
+            in_array(ClassUtils::getClass($entity), $this->versionableEntities)
         ) {
             $pendings[] = $entity;
         }